backend/requirements.txt
# Python backend dependencies
mistralai>=1.0.0
chromadb>=0.5.0
langchain-core>=0.2.0
langgraph>=0.1.0
langchain-text-splitters>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0
pdfplumber>=0.11.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0